lift forEachADCoord into llama
